Collections Structures, Collections Operators/Collections Methods in Flutter/Dart
We will discuss iterable, list, map and set again and see various looping techniques we can do on them. Yes, we've seen list set and map before. Now we will see what an iterable is, how these are related to iterable, and how it allows for in looping. We will see collection operators and the if and for commands that we can use in collections. We will talk about functions that make transformations on collections, take functions as parameters and apply this function to the elements of the collection.